requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autocuda==0.11
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.7
brotlipy==0.7.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.2.1
en-core-web-lg==3.2.0
en-core-web-sm==3.2.0
entrypoints==0.4
executing==0.8.3
fastapi==0.76.0
fastjsonschema==2.15.3
filelock==3.6.0
findfile==1.7.8.1
flit_core==3.7.1
fonttools==4.32.0
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
jupyterthemes==0.20.0
kiwisolver==1.4.2
langcodes==3.3.0
langdetect==1.0.9
lesscpy==0.15.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
metric-visualizer==0.4.21
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
murmurhash==1.0.6
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
nexus-ai==1.0
nltk==3.7
notebook==6.4.10
numpy==1.22.3
orjson==3.6.8
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.4
ply==3.11
pmdarima==1.8.5
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pyabsa==1.14.4
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytorch-warmup==0.1.0
pytz==2022.1
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.5
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools==58.0.4
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
stack-data==0.2.0
starlette==0.18.0
statsmodels==0.13.2
tabulate==0.8.9
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thinc==8.0.15
threadpoolctl==3.1.0
tikzplotlib==0.10.1
tinycss2==1.1.1
tokenizers==0.11.6
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.1.1
ujson==5.2.0
update-checker==0.18.0
urllib3==1.26.8
uvicorn==0.17.6
wasabi==0.9.1
watchgod==0.8.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websockets==10.3
wheel==0.37.1
widgetsnbextension==3.6.0
win-inet-pton==1.1.0
wincertstore==0.2
wsproto==1.1.0
zipp==3.8.0
